build/windows: Port building scripts to use Windows native shell
From now, Windows contributors can use the default shell provided by the OS
(which is PS since Windows build 10.0.14971.0), like Linux and macOS users do.
We still use MSYS2 but not the POSIX shell. This change adds these features:

'1_build-deps-msys2.sh' is now '1_build-deps-msys2.ps1'
- Faster clonning by using 'git-scm' (the MSYS2 one have performance problems)
- Easier to use non-MSYS2 binaries, not only 'git-scm' but also official meson,
  deps from vckpg etc. This is a needed step towards the future use of MSVC.

'2_build-gimp-msys2.sh' is now '2_build-gimp-msys2.ps1'
- By default, vanilla builds (normally triggered on PS) will create a bundle,
  dropping the need of 'gimp.cmd' wrapper and .typelib and .interp limitations
  Builds configured to be non-relocatable still work on MSYS2 shells, as always.
  This change is important because even Windows devs more experienced than me
  get confused with all the relocatbility stuff, which is the default on Windows.

'2_bundle-gimp-uni_base.sh'
- As a result of the change above, bundling code was changed to be a bit faster.

Since we aren't not using a POSIX shell ('mintty') anymore:
- GIMP can be built from every location easily with Righ Click "Open Terminal",
  IDE integrations etc, without needing to manual tweak MSYS2 .ini files etc.
  We could tweak MSYS2 to get some features above but not top-tier integration.
- Developers can be more aware of Windows native vars, paths etc, and avoid bugs
  Some build files were improved to support the 'Windows way of doing things'.
- No need to close and reopen anything anymore after running 'pacman -Suy'!

+++ build/windows/2_bundle-gimp-uni_base.sh
@@ -2,27 +2,14 @@
 
 set -e
 
-if [ "$1" != '--authorized' ] && [ "$1" != '--force' ]; then
-  # We can't easily figure out if GIMP was built relocatable so
-  # let's prevent contributors from creating broken bundles
-  echo -e "\033[31m(ERROR)\033[0m: Script called standalone. Please, run GIMP build script with '--relocatable' or this bundling script with '--force'"
-  exit 1
-fi
-
-if [ "$1" = '--force' ] && [ "$MSYSTEM_CARCH" = "i686" ]; then
-  echo -e "\033[33m(WARNING)\033[0m: 32-bit builds will be dropped in a future release. See: https://gitlab.gnome.org/GNOME/gimp/-/issues/10922"
-fi
-
-
-# NOTE: The bundling scripts, different from building scripts, need to set
-# the ARTIFACTS_SUFFIX, even locally: 1) to avoid confusion (bundle dirs are
-# relocatable so can be copied to a machine with other arch); and 2) to our
+# NOTE: The bundling scripts need to set the ARTIFACTS_SUFFIX: to our
 # dist scripts fallback code be able to identify what they are distributing
-if [ "$MSYSTEM_CARCH" = "aarch64" ]; then
+UNIX_ARCH=$(uname -m)
+if [ "$UNIX_ARCH" = "aarch64" ]; then
   export ARTIFACTS_SUFFIX="-a64"
-elif [ "$CI_JOB_NAME" = "gimp-win-x64-cross" ] || [ "$MSYSTEM_CARCH" = "x86_64" ]; then
+elif [ "$CI_JOB_NAME" = "gimp-win-x64-cross" ] || [ "$UNIX_ARCH" = "x86_64" ]; then
   export ARTIFACTS_SUFFIX="-x64"
-else # [ "$MSYSTEM_CARCH" = "i686" ];
+else # [ "$UNIX_ARCH" = "i686" ];
   export ARTIFACTS_SUFFIX="-x86"
 fi
 
@@ -39,60 +26,57 @@ fi
 
 
 # Bundle deps and GIMP files
+## When this script is called from POSIX shell
 if [[ "$CI_JOB_NAME" =~ "cross" ]]; then
   export GIMP_PREFIX="`realpath ./_install-cross`"
   export MSYS_PREFIX="$GIMP_PREFIX"
-else
-  if [ "$GITLAB_CI" ]; then
-    export GIMP_PREFIX="$PWD/_install"
-  elif [ -z "$GITLAB_CI" ] && [ -z "$GIMP_PREFIX" ]; then
-    export GIMP_PREFIX="$PWD/../_install"
-  fi
-  export MSYS_PREFIX="$MSYSTEM_PREFIX"
 fi
+export GIMP_PREFIX="$PWD/../_install"
+export MSYS_PREFIX="$MSYSTEM_PREFIX"
 export GIMP_DISTRIB="`realpath ./gimp`${ARTIFACTS_SUFFIX}"
+## When this script is called from Windows shell
+if [ -z "$MSYSTEM" ]; then
+  export GIMP_PREFIX="$MESON_INSTALL_DESTDIR_PREFIX"
+  export MSYS_PREFIX=$(grep '    directx-sdk-dir' $MESON_BUILD_ROOT/meson-logs/meson-log.txt | sed 's/    directx-sdk-dir   : //')
+  if [ -z "$GITLAB_CI" ]; then
+    export GIMP_DISTRIB="$GIMP_PREFIX"
+  fi
+  export GIMP_SOURCE=$(echo $MESON_SOURCE_ROOT/ | sed 's|\\|/|g')
+fi
 
 bundle ()
 {
-  check_recurse=${2##*/}
-  # Copy files with wildcards
-  if [[ "$check_recurse" =~ '*' ]] && [[ ! "$check_recurse" =~ '/' ]]; then
-    path_dest_parent=$(echo $1/${2%/*} | sed "s|${1}/||")
-    mkdir -p "$GIMP_DISTRIB/$path_dest_parent"
-    bundledArray=($(find $1/${2%/*} -maxdepth 1 -name ${2##*/}))
-    for path_origin_full2 in "${bundledArray[@]}"; do
-      echo "(INFO): copying $path_origin_full2 to $GIMP_DISTRIB/$path_dest_parent"
-      cp $path_origin_full2 "$GIMP_DISTRIB/$path_dest_parent"
-    done
-  # Copy specific file or specific folder
-  else
-    path_origin_full=$(echo $1/$2)
-    if [[ "$2" =~ '/' ]]; then
-      path_dest_parent=$(dirname $path_origin_full | sed "s|${1}/||")
-    fi
-    if [ -d "$path_origin_full" ] || [ -f "$path_origin_full" ]; then
-      mkdir -p "$GIMP_DISTRIB/$path_dest_parent"
-      echo "(INFO): copying $path_origin_full to $GIMP_DISTRIB/$path_dest_parent"
-      cp -r "$path_origin_full" "$GIMP_DISTRIB/$path_dest_parent"
-    else
-      echo -e "\033[33m(WARNING)\033[0m: $path_origin_full does not exist!"
-    fi
+  search_path=$(dirname $(echo $1/$2))
+  if [ ! -d "$search_path" ]; then
+    search_path=$(echo $1/${2%%/*})
   fi
+  bundledArray=($(find "$search_path" -maxdepth 1 -name ${2##*/}))
+  for target_path in "${bundledArray[@]}"; do
+    bundle_path=$(echo $target_path | sed "s|$1/|$GIMP_DISTRIB/|g")
+    parent_path=$(dirname $bundle_path)
+    if [ ! "$MSYSTEM" ] && [ "$1" != "$GIMP_PREFIX" ]; then
+      echo "Bundling $target_path to $parent_path"
+    fi
+    mkdir -p "$parent_path"
+    cp -fru "$target_path" $parent_path >/dev/null 2>&1 || continue
+  done
 }
 
 clean ()
 {
-  if [[ "$2" =~ '/' ]]; then
-    cleanedArray=($(find $1/${2%/*} -iname ${2##*/}))
-  else
-    cleanedArray=($(find $1/ -iname ${2##*/}))
-  fi
-  for path_dest_full in "${cleanedArray[@]}"; do
-    if [[ "$path_dest_full" = "${cleanedArray[0]}" ]]; then
-      echo "(INFO): cleaning $1/$2"
+  if [ "$INSTALLER_OPTION" ] || [ "$STORE_OPTION" ]; then
+    if [[ "$2" =~ '/' ]]; then
+      cleanedArray=($(find $1/${2%/*} -iname ${2##*/}))
+    else
+      cleanedArray=($(find $1/ -iname ${2##*/}))
     fi
-    rm $path_dest_full
-  done
+    for path_dest_full in "${cleanedArray[@]}"; do
+      if [[ "$path_dest_full" = "${cleanedArray[0]}" ]]; then
+        echo "Cleaning $1/$2"
+      fi
+      rm $path_dest_full
+    done
+  fi
 }
 
 
@@ -100,47 +84,41 @@ clean ()
 mkdir -p $GIMP_DISTRIB
 echo "*" > $GIMP_DISTRIB/.gitignore
 
-## Copy a previously built wrapper at tree root, less messy than
-## having to look inside bin/, in the middle of all the DLLs.
-## This also configure the interpreters for local builds as courtesy.
-bundle "$GIMP_PREFIX" gimp.cmd
-
-
 ## Settings.
-bundle "$GIMP_PREFIX" etc/gimp/
+bundle "$GIMP_PREFIX" etc/gimp
 ### Needed for fontconfig
-bundle "$MSYS_PREFIX" etc/fonts/
+bundle "$MSYS_PREFIX" etc/fonts
 
 
 ## Headers (in evaluation): https://gitlab.gnome.org/GNOME/gimp/-/issues/6378.
-#bundle $GIMP_PREFIX/include/gimp-*/
-#bundle $GIMP_PREFIX/include/babl-*/
-#bundle $GIMP_PREFIX/include/gegl-*/
+#bundle $GIMP_PREFIX/include/gimp-*
+#bundle $GIMP_PREFIX/include/babl-*
+#bundle $GIMP_PREFIX/include/gegl-*
 
 
 ## Library data.
-bundle "$GIMP_PREFIX" lib/gimp/
-bundle "$GIMP_PREFIX" lib/babl-*/
-bundle "$GIMP_PREFIX" lib/gegl-*/
-bundle "$MSYS_PREFIX" lib/gio/
+bundle "$GIMP_PREFIX" lib/gimp
+bundle "$GIMP_PREFIX" lib/babl-*
+bundle "$GIMP_PREFIX" lib/gegl-*
+bundle "$MSYS_PREFIX" lib/gio
 bundle "$MSYS_PREFIX" lib/gdk-pixbuf-*/*/loaders.cache
 bundle "$MSYS_PREFIX" lib/gdk-pixbuf-*/*/loaders/libpixbufloader-png.dll
-bundle "$MSYS_PREFIX" lib/gdk-pixbuf-*/*/loaders/libpixbufloader-svg.dll
+bundle "$MSYS_PREFIX" lib/gdk-pixbuf-*/*/loaders/pixbufloader_svg.dll
 clean "$GIMP_DISTRIB" lib/*.a
 
 
 ## Resources.
-bundle "$GIMP_PREFIX" share/gimp/
+bundle "$GIMP_PREFIX" share/gimp
 ### Needed for file dialogs
 bundle "$MSYS_PREFIX" share/glib-*/schemas/gschemas.compiled
 ### https://gitlab.gnome.org/GNOME/gimp/-/issues/6165
-bundle "$MSYS_PREFIX" share/icons/Adwaita/
+bundle "$MSYS_PREFIX" share/icons/Adwaita
 ### https://gitlab.gnome.org/GNOME/gimp/-/issues/5080
-bundle "$GIMP_PREFIX" share/icons/hicolor/
+bundle "$GIMP_PREFIX" share/icons/hicolor
 ### Needed for file-wmf work
-bundle "$MSYS_PREFIX" share/libwmf/
+bundle "$MSYS_PREFIX" share/libwmf
 ### Only copy from langs supported in GIMP.
-lang_array=($(echo $(ls po/*.po |
+lang_array=($(echo $(ls ${GIMP_SOURCE}po/*.po |
               sed -e 's|po/||g' -e 's|.po||g' | sort) |
               tr '\n\r' ' '))
 for lang in "${lang_array[@]}"; do
@@ -154,9 +132,9 @@ done
 ### Needed for welcome page
 bundle "$GIMP_PREFIX" share/metainfo/org.gimp*.xml
 ### mypaint brushes
-bundle "$MSYS_PREFIX" share/mypaint-data/
+bundle "$MSYS_PREFIX" share/mypaint-data
 ### Needed for full CJK and Cyrillic support in file-pdf
-bundle "$MSYS_PREFIX" share/poppler/
+bundle "$MSYS_PREFIX" share/poppler
 
 
 ## Executables and DLLs.
@@ -175,8 +153,8 @@ bundle "$MSYS_PREFIX" bin/gdbus.exe
 
 ### Optional binaries for GObject Introspection support
 if [ "$CI_JOB_NAME" != 'gimp-win-x64-cross' ]; then
-  bundle "$GIMP_PREFIX" lib/girepository-*/
-  bundle "$MSYS_PREFIX" lib/girepository-*/
+  bundle "$GIMP_PREFIX" lib/girepository-*
+  bundle "$MSYS_PREFIX" lib/girepository-*
 
   # https://gitlab.gnome.org/GNOME/gimp/-/issues/11597
   #bundle "$MSYS_PREFIX" bin/luajit.exe
@@ -184,7 +162,7 @@ if [ "$CI_JOB_NAME" != 'gimp-win-x64-cross' ]; then
   #bundle "$MSYS_PREFIX" share/lua/
 
   bundle "$MSYS_PREFIX" bin/python*.exe
-  bundle "$MSYS_PREFIX" lib/python*/
+  bundle "$MSYS_PREFIX" lib/python*
   clean "$GIMP_DISTRIB" lib/python*/*.pyc
 else
   # Just to ensure there is no introspected files that will output annoying warnings
@@ -196,15 +174,15 @@ else
 fi
 
 ### Deps (DLLs) of the binaries in 'lib' and 'bin' dirs
-echo "(INFO): searching for dependencies of $GIMP_DISTRIB/lib in $GIMP_PREFIX and $MSYS_PREFIX"
+echo "Searching for dependencies of $GIMP_DISTRIB/lib in $GIMP_PREFIX and $MSYS_PREFIX"
 libArray=($(find "$GIMP_DISTRIB/lib" \( -iname '*.dll' -or -iname '*.exe' \)))
 for dep in "${libArray[@]}"; do
-  python3 build/windows/2_bundle-gimp-uni_dep.py $dep $GIMP_PREFIX/ $MSYS_PREFIX/ $GIMP_DISTRIB --output-dll-list done-dll.list;
+  python3 ${GIMP_SOURCE}build/windows/2_bundle-gimp-uni_dep.py $dep $GIMP_PREFIX/ $MSYS_PREFIX/ $GIMP_DISTRIB --output-dll-list done-dll.list;
 done
-echo "(INFO): searching for dependencies of $GIMP_DISTRIB/bin in $GIMP_PREFIX and $MSYS_PREFIX"
+echo "Searching for dependencies of $GIMP_DISTRIB/bin in $GIMP_PREFIX and $MSYS_PREFIX"
 binArray=($(find "$GIMP_DISTRIB/bin" \( -iname '*.dll' -or -iname '*.exe' \)))
 for dep in "${binArray[@]}"; do
-  python3 build/windows/2_bundle-gimp-uni_dep.py $dep $GIMP_PREFIX/ $MSYS_PREFIX/ $GIMP_DISTRIB --output-dll-list done-dll.list;
+  python3 ${GIMP_SOURCE}build/windows/2_bundle-gimp-uni_dep.py $dep $GIMP_PREFIX/ $MSYS_PREFIX/ $GIMP_DISTRIB --output-dll-list done-dll.list;
 done
 
 ### .pdb (CodeView) debug symbols
@@ -212,8 +190,3 @@ done
 #if [ "$CI_JOB_NAME" != "gimp-win-x64-cross" ]; then
 #  cp -fr ${GIMP_PREFIX}/bin/*.pdb ${GIMP_DISTRIB}/bin/
 #fi
-
-
-# Delete wrapper to prevent contributors from running a
-# relocatable build expecting it to work non-relocatable
-rm -r $GIMP_PREFIX/gimp.cmd
